#63 [REFACTOR] mypage-review 역할별 컴포넌트 분리
